refactor : numbergenerator 패키지 생성 (관련 객체 이관)
