+++ src/main/java/lotto/LottoMain.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import lotto.domain.*;
+import lotto.domain.numbergenerator.LottoNumberGeneratorFactory;
 import lotto.view.InputView;
 import lotto.view.OutputView;
