+++ src/main/java/lotto/domain/numbergenerator/LottoNumberGeneratorFactory.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.numbergenerator;
+
+import lotto.domain.LottoBuyDetails;
 
 public class LottoNumberGeneratorFactory {
     private LottoBuyDetails lottoBuyDetails;
